requirements.txt
requests
whatismyip
dnspython
geocoder
Flask
Gunicorn
python-whois
whois
pyOpenSSL
bcrypt
flask_mysqldb
flask_session
python-dotenv
pandas
flask_mail